+++ mafComparator/src/cString.c
@@ -92,8 +92,8 @@ int cStr_startsWith(char *string, char *query, int ignorecase) {
 	}
 }
 
-int32_t cStr_getIntLength(int32_t n) {
-	int32_t count = 0;
+int64_t cStr_getIntLength(int64_t n) {
+	int64_t count = 0;
 	do {
 		count++;
 	} while ((n /= 10) > 0);
@@ -138,8 +138,8 @@ void cStr_appendChar(char *s, char c) {
 	s[len + 1] = '\0';
 }
 
-char *cStr_getStringFromIntArray(int32_t *array, int32_t size, const char sep) {
-	int32_t i;
+char *cStr_getStringFromIntArray(int64_t *array, int64_t size, const char sep) {
+	int64_t i;
 	int numChars = 0;
 	char *string = NULL;
 	char buffer[64];
